% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{cog2015}
\alias{cog2015}
\title{cog2015}
\format{An object of class \code{data.frame} with 519334 rows and 2048 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
cog2015
}
\description{
Questionnaire About Various School Topics
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
