% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{cps2015}
\alias{cps2015}
\title{cps2015}
\format{An object of class \code{data.frame} with 414498 rows and 190 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
cps2015
}
\description{
Results of Students Ability to Work Together to Solve a Particular Problem
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
