% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R, R/DatasetsDocVars.R
\docType{data}
\name{Data}
\alias{Data}
\alias{cog2015.part1}
\alias{cog2015.part2}
\alias{cog2015.part3}
\alias{cog2015.part4}
\alias{cps2015}
\alias{lit2015}
\alias{sch2015}
\alias{stud2015.part1}
\alias{stud2015.part2}
\alias{stud2015_2}
\alias{tch2015}
\alias{tim2015}
\title{PISA 2015 Datasets}
\format{Dataframe with the Specific Variables for the Dataset
A more detailed description can be found in the corresponding pdf file}
\usage{
cog2015.part1

cog2015.part2

cog2015.part3

cog2015.part4

cps2015

lit2015

sch2015

stud2015.part1

stud2015.part2

stud2015_2

tch2015

tim2015

cog2015.part2

cog2015.part3

cog2015.part4

cps2015

lit2015

sch2015

stud2015.part1

stud2015.part2

stud2015_2

tch2015

tim2015
}
\description{
These are the datasets from the PISA 2015 Survey
}
\keyword{datasets}
