% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESEV552.R
\docType{data}
\name{DataVariables}
\alias{DataVariables}
\alias{stud2015vars}
\alias{stud2015_2vars}
\alias{cog2015vars}
\alias{sch2015vars}
\alias{tim2015vars}
\alias{tch2015vars}
\alias{lit2015vars}
\alias{cps2015vars}
\title{Description of Dataset Variables}
\format{A data frame of 3 Variables
\describe{
\item{Names}{The Name of the Variable}
\item{Description}{A brief Description of the Variable}
\item{Type}{The Type of the Variable (Nominal, Ordinal, Numeric, Ratio)}
}}
\source{
\url{"http://www.oecd.org/pisa/"}
}
\usage{
stud2015vars

stud2015_2vars

cog2015vars

sch2015vars

tim2015vars

tch2015vars

lit2015vars

cps2015vars
}
\description{
These are the descriptions of each of the variables for a specific dataset from the PISA 2015 Survey
}
\keyword{datasets}
