% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDocVars.R
\docType{data}
\name{DataVars}
\alias{DataVars}
\alias{cog2015.part1}
\title{PISA 2015 Dataset Variables}
\format{Dataframe with the Specific Variables for the Dataset
\describe{
 \item {Names}{Name of the Variable from the Dataset}
 \item {Description}{Brief Description of Variable from the Dataset}
 \item {Type}{Type of Variable}
}}
\usage{
cog2015.part1
}
\description{
These are the Variable Description Datasets from the PISA 2015 Survey
}
\keyword{datasets}
