% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{lit2015}
\alias{lit2015}
\title{lit2015}
\format{An object of class \code{data.frame} with 123041 rows and 217 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
lit2015
}
\description{
Survey of Students regarding Money and a Questionnaire on Application of their Knowledge
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
