% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDocVars.R
\docType{data}
\name{lit2015vars}
\alias{lit2015vars}
\title{lit2015vars}
\format{A dataframe of 3 Variables
\describe{
 \item{Names}{Name of the Variable}
 \item{Description}{Brief Description of the Variable}
 \item{Type}{What Type of Varible is it? (Nominal, Ordinal, Numeric, Ratio)}
}}
\usage{
lit2015vars
}
\description{
Dataset describing the variables for lit2015
}
\keyword{datasets}
