% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESEV552.R
\name{PISAData}
\alias{PISAData}
\title{PISA Data}
\usage{
PISAData()
}
\value{
The eight datasets into your personal workspace
}
\description{
This is a function to download the PISA 2015 Data from the "mstuart2097/RESEV552Data" repository in Github
}
