% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESEV552.R
\name{PISAShinyApp}
\alias{PISAShinyApp}
\title{PISAShinyApp}
\format{Three Shiny Outputs about the Variable Choice for a Specific Dataset
\describe{
\item{Description}{A Brief Description of the Chosen Variable}
\item{Table}{A Table of the Responses for the Chosen Variable (Summary of the Repsonses if the Variable is Numeric)}
\item{Plot}{A Bar Plot of the 10 Most Popular Repsonses for the Chosen Variable (Histogram if the Variable is Numeric)}
\item{Plot}{A Bar Plot of the 10 Most Popular Repsonses for the Chosen Variable (Histogram if the Variable is Numeric)}
}}
\usage{
PISAShinyApp()
}
\value{
A Shiny App with choices of Dataset, Type of Variable and Specific Variable
}
\description{
This is a Shiny App which describes all of the PISA 2015 datasets
}
