% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{sch2015}
\alias{sch2015}
\title{sch2015}
\format{An object of class \code{data.frame} with 17908 rows and 273 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
sch2015
}
\description{
Survey of School Administrators regarding School Resources and Teacher Interaction 
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
