% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{stud2015}
\alias{stud2015}
\title{stud2015}
\format{An object of class \code{data.frame} with 519334 rows and 921 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
stud2015
}
\description{
Survey of Students and Their Parents/Guardians about General Questions of Schooling, Study Habits, and Access to Study Materials 
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
