% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{stud2015_2}
\alias{stud2015_2}
\title{stud2015_2}
\format{An object of class \code{data.frame} with 519334 rows and 21 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
stud2015_2
}
\description{
Survey of Students regarding Bullying and Self-Esteem 
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
