% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDocVars.R
\docType{data}
\name{stud2015_2vars}
\alias{stud2015_2vars}
\title{stud2015_2vars}
\format{A dataframe of 3 Variables
\describe{
 \item{Names}{Name of the Variable}
 \item{Description}{Brief Description of the Variable}
 \item{Type}{What Type of Varible is it? (Nominal, Ordinal, Numeric, Ratio)}
}}
\usage{
stud2015_2vars
}
\description{
Dataset describing the variables for stud2015_2
}
\keyword{datasets}
