% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{tch2015}
\alias{tch2015}
\title{tch2015}
\format{An object of class \code{data.frame} with 108292 rows and 257 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
tch2015
}
\description{
Survey of Teachers regarding Their Training and Current Job 
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
