% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsDoc.R
\docType{data}
\name{tim2015}
\alias{tim2015}
\title{tim2015}
\format{An object of class \code{data.frame} with 447400 rows and 131 columns.}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
tim2015
}
\description{
Time Taken to Answers Specific Questions from stud2015 and stud2015_2
A brief description of each of the variables can be found in the corresponding "vars" file
}
\keyword{datasets}
