% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESEV552.R
\name{PISAData}
\alias{PISAData}
\title{PISA Data}
\usage{
PISAData(dataset)
}
\arguments{
\item{dataset}{The dataset you want to download from the "RESEV552Data" repository}
}
\value{
The specified dataset into your personal workspace
}
\description{
This is a function to download a PISA 2015 Dataset from the "mstuart2097/RESEV552Data" repository in Github
}
