% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{stud2015}
\alias{stud2015}
\title{PISA2015}
\format{A dataframe of the survey results for each respondent
A description of each of the dataframes, including important variables, can be found in PISA2015_Datasets_Description.pdf
A brief description of each of the variables can be found in the corresponding "vars" file}
\source{
\url{"http://www.oecd.org/pisa/data/2015database/"}
}
\usage{
stud2015
}
\description{
These are the datasets from the PISA 2015 Survey to be Used in this class
}
\keyword{datasets}
