% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsVars.R
\docType{data}
\name{stud2015vars}
\alias{stud2015vars}
\title{PISA2015vars}
\format{A dataframe of 3 Variables
\describe{
 \item{Names}{Name of the Variable}
 \item{Description}{Brief Description of the Variable}
 \item{Type}{What Type of Varible is it? (Nominal, Ordinal, Numeric, Ratio)}
}}
\usage{
stud2015vars
}
\description{
These are the datasets describing the variables for the corresponding dataset from the PISA 2015 Survey
}
\keyword{datasets}
